:building_construction: Separate Revisions from Core Learning Object Logic
This commit takes the ability to read and create revisions and moves it
into a subdirectory of LearningObjects, with its own interactor and
RouteHandler. This gives the RESTful Resource a more distinctive home in
the project and better adheres to SRP (changes to Revision logic won't
require us to muck around with other Learning Object functionality).

Additionally, tests for the LearningObjectAdapter that were testing
Revision logic have been converted to interactor-level tests in the
migration, since the purpose of testing the adapter was in fact to
validate that the interactor worked as expected.
