+++ src/LearningObjects/Revisions/interactors/index.ts
@@ -1,0 +1,4 @@
+import { getLearningObjectRevision } from './getLearningObjectRevision';
+import { createLearningObjectRevision } from './createLearningObjectRevision';
+
+export { getLearningObjectRevision, createLearningObjectRevision };
